Updated the rendering system to shade trees better, added scale and frequency to compiler macros, flooding is activated but the spill number is set to zero (effectively off - not fully satisfied with the water table computation system), reduced accuracy to floats becaues it works and makes it faster

+++ source/vegetation.h
@@ -16,7 +16,7 @@ struct Plant{
   const float rate = 0.05;
 
   void grow();
-  void root(double* density, glm::ivec2 dim, double factor);
+  void root(float* density, glm::ivec2 dim, float factor);
 
   Plant& operator=(const Plant& o){
     if(this != &o){  //Self Check
@@ -33,7 +33,7 @@ void Plant::grow(){
   size += rate*(maxsize-size);
 };
 
-void Plant::root(double* density, glm::ivec2 dim, double f){
+void Plant::root(float* density, glm::ivec2 dim, float f){
 
   //Can always do this one
   density[index]       += f*1.0;
